backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "bson>=0.5.10",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.0.0",
    "mcp[cli]>=1.20.0",
    "metaflow>=2.19.5",
    "motor>=3.7.1",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.4",
    "pymongo>=4.15.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "python-socketio>=5.14.3",
    "scipy>=1.16.3",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
]
